pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "django>=5.2",
    "django-allauth[socialaccount]>=65.7.0",
    "django-browser-reload>=1.18.0",
]
